// src/app/models/foto.ts
export class Foto {
    id: string = "";
    rg: string = "";
    cpf: string = "";
    nome: string = "";
    vulgo: string = "";
    mae: string = "";
    bairro: string = "";
    area: string = "";

    pai: string = "";
    nascimento: string = "";
    logradouro: string = "";
    bairro_id: string = "";
    atuacao_id: string = "";

}